kayttaja api pathin siirto omaan routeriin
